prefetching: remove prefetched item after use in udf (#818)
* prefetching: remove prefetched item after use in udf

This PR removes the prefetched item after use in the UDF.
This is enabled by default on `prefetch>0`, unless `cache=True` is set in the UDF, in
which case the prefetched item is not removed.

For pytorch dataloader, this is not enabled by default, but can be enabled by setting
`remove_prefetched=True` in the `PytorchDataset` class.
This is done so because the dataset can be used in multiple epochs, and removing the
prefetched item after use can cause it to redownload again in the next epoch.

The exposed `remove_prefetched=True|False` setting could be renamed to
some better option. Feedbacks are welcome.

* close iterable properly