e2e tests: limit name_len_slow to 3, split e2e tests from other tests (#842)
[The `test_query_e2e` takes almost ~8mins to run][1] (whole CI job takes
11 mins). The `name_len_slow` script is the main culprit, since it
sleeps for 1 sec in each udf function and that mapper is run in a single
process parallel mode.

```
474.21s call     tests/test_query_e2e.py::test_query_e2e@tmpfile
```

This commit adds a limit of 3 files to the name_len_slow script, which is
enough, since it's only running a single process.
(We immediately interrupt the running process after seeing "UDF
Processing Started" gets printed).

This also split tests into two: one for the e2e tests and one for the
rest, so that these things are more obvious in the future.

[1]: https://github.com/iterative/datachain/actions/runs/12879531971/job/35907168617#step:8:82