pull _is_file checks to get_listing
The parsing utility should not hit remote servers at all.
Also returning `None` does not make sense for `parse_listing_uri`
when the uri is a file since there is no listing.

It's better to keep this function naive than to add other logic into
it. And I think `get_listing` is a better place for that, to have all
kinds of logic with side-effects.

The only other place where this function is used besides `get_listing` is in `DatasetDependency.dataset_name`.
In that case, a dataset already exists and since there can be no listing
for a file uri, it's safe to assume that the uri is either a glob or a
directory path.

https://github.com/iterative/datachain/blob/e45a62f43cad47a162345f01e3a2f9a1fff8f47a/src/datachain/dataset.py#L94

I stumbled upon this when I was investigating why a unit-test was slow.
Turns out it was hitting the remote with a fake s3 path for `is_file()`,
which was failing and taking 4-5 seconds to run.

https://github.com/iterative/datachain/blob/e45a62f43cad47a162345f01e3a2f9a1fff8f47a/tests/unit/test_dataset.py#L88-L99