fix(huggingface): wrap HfFileSystem to an async filesystem
This commit wraps the HfFileSystem class to convert to an async filesystem.
This is done by using the `AsyncFileSystemWrapper` class from the `fsspec` library, which requires `fsspec>=2024.12.0`.

With the wrapper, all the functions will run in a separate thread in an
executor, and won't block the main event-loop.

This will now support all the async APIs that the `Client` expects, and
will hence support caching of the files.

The alternative here is to have two different implementations, one async
and another sync, but that is too much of an effort at this time.

Fixes #661.