fix(huggingface): wrap HfFileSystem to an async filesystem (#957)
* fix(huggingface): wrap HfFileSystem to an async filesystem

This commit wraps the HfFileSystem class to convert to an async filesystem.
This is done by using the `AsyncFileSystemWrapper` class from the `fsspec` library, which requires `fsspec>=2024.12.0`.

With the wrapper, all the functions will run in a separate thread in an
executor, and won't block the main event-loop.

This will now support all the async APIs that the `Client` expects, and
will hence support caching of the files.

The alternative here is to have two different implementations, one async
and another sync, but that is too much of an effort at this time.

Fixes #661.

* version_aware does not always exist

* set protocol to the wrapper class, AsyncFileSystemWrapper does not seem to do that

* ci: run llm_and_nlp test only on a single job

Other jobs were noop, as token was not being set.

* hf: create filesystem class lazily, raise error on fsspec<2024.12.0