fix(tests): disable coverage in test_import_time (#975)
* fix(tests): disable coverage in test_import_time

This commit disables coverage for the test_import_time test to avoid
unpredictable results. The import usually takes  ~400ms but with
coverage enabled, it was frequently going past 600ms on my machine.

Which i believe is also happening on the CI.

Also added a log message to display the full contents of the output
from `-Ximporttime` in case of parse error in `from_csv`.

* Update tests/test_import_time.py

* do not generate coverage result on e2e tests

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>