POC: Add an inexpensive test to set anon if no creds are found in gcs
Regarding the discussion on #740, this is a possible approach to specify
anon token in gcs if no creds is found.
This is the similar to the approach we already have s3.

The approach is to try different method of connection before defaulting
back to anon as gcs filesystem does. But we are skipping the metadata
check for the case of google cloud engine.

Testing the improvement with the code:
```py
import time
from datachain import DataChain

start = time.time()
images = DataChain.from_storage("gs://datachain-demo/dogs-and-cats/*jpg")
images.show()
end = time.time()
print(f"Time taken: {end - start} seconds")
```

The improvement was from:
27.73566699028015 seconds

to:
Time taken: 6.101557016372681 seconds

Closes #740