read_database/read_records: set default for the column instead of null (#1049)
This is a partial revert of e201aefaf9 from #1034, which added support
for inferring schema with `Optional[T]` types, and also added support
for creating nullable columns in the database.

Also, this commit sets the default value for a given column if the value
is `None` which depends on a given dialect/database. For example, for
SQLite, the default value for a `String` column is `None`, while for ClickHouse,
it is an empty string.

Support for nullable columns can be added separately in the future.