Save team name to global by default
When setting the team, the team name is set locally by default. Similarly, when logging in, the token is saved globally by default.
To unify this behavior, make it so that team name is also saved to global by default allowing user to save it locally if they want using local flag.