Introduce read_listing_dataset for optimized reading of listing dataset (#1150)
* Introduce read_listing_dataset for optimized reading of listing dataset

For the case of listing dataset, that starts with lst__ , we have fixed set of the columns and listing. For the same reason, we can skip the autoload with the engine that fetches the metadata from the warehouse. This skips a few connection to the warehouse and also it skips the whole get_table part to get the tables from the warehouse.

Taken from https://github.com/iterative/studio/pull/11694

Relates to https://github.com/iterative/studio/issues/11704

* Fix the PR comments

* Cover missing coverage